toString.latextoken {latex2exp}	R Documentation

Converts a token created by TeX() to a string, later to be parsed into an expression (for internal use).

Description

Converts a token created by TeX() to a string, later to be parsed into an expression (for internal use).

Usage

## S3 method for class 'latextoken'
toString(x, ...)

Arguments

x
The TeX() token

...
Additional arguments (ignored)

Value

A string

[Package latex2exp version 0.4.0 Index]